+++ examples/budget/main.tf
@@ -1,7 +1,7 @@
 // Copyright (c) 2017, 2019, Oracle and/or its affiliates. All rights reserved.
 
 /*
- * This example shows how to use the budget resource.
+ * This example shows how to use the budget and alert rule resources.
  */
 
 variable "tenancy_ocid" {}
@@ -37,25 +37,24 @@ data "oci_budget_budget" "budget1" {
 }
 
 output "budget" {
-  value = <<EOF
+  value = {
+    amount                = "${data.oci_budget_budget.budget1.amount}"
+    compartment_id        = "${data.oci_budget_budget.budget1.compartment_id}"
+    reset_period          = "${data.oci_budget_budget.budget1.reset_period}"
+    target_compartment_id = "${data.oci_budget_budget.budget1.target_compartment_id}"
+    description           = "${data.oci_budget_budget.budget1.description}"
+    display_name          = "${data.oci_budget_budget.budget1.display_name}"
+    alert_rule_count      = "${data.oci_budget_budget.budget1.alert_rule_count}"
+    state                 = "${data.oci_budget_budget.budget1.state}"
+    time_created          = "${data.oci_budget_budget.budget1.time_created}"
+    time_updated          = "${data.oci_budget_budget.budget1.time_updated}"
+    version               = "${data.oci_budget_budget.budget1.version}"
 
-  amount = "${data.oci_budget_budget.budget1.amount}"
-  compartment_id = "${data.oci_budget_budget.budget1.compartment_id}"
-  reset_period = "${data.oci_budget_budget.budget1.reset_period}"
-  target_compartment_id = "${data.oci_budget_budget.budget1.target_compartment_id}"
-  description = "${data.oci_budget_budget.budget1.description}"
-  display_name = "${data.oci_budget_budget.budget1.display_name}"
-  alert_rule_count = "${data.oci_budget_budget.budget1.alert_rule_count}"
-  state = "${data.oci_budget_budget.budget1.state}"
-  time_created = "${data.oci_budget_budget.budget1.time_created}"
-  time_updated = "${data.oci_budget_budget.budget1.time_updated}"
-  version = "${data.oci_budget_budget.budget1.version}"
-EOF
-
-  # These values are not always present
-  //  actual_spend = "${data.oci_budget_budget.budget1.actual_spend}"
-  //  forecasted_spend = "${data.oci_budget_budget.budget1.forecasted_spend}"
-  //  time_spend_computed = "${data.oci_budget_budget.budget1.time_spend_computed}"
+    # These values are not always present
+    //    actual_spend        = "${data.oci_budget_budget.budget1.actual_spend}"
+    //    forecasted_spend    = "${data.oci_budget_budget.budget1.forecasted_spend}"
+    //    time_spend_computed = "${data.oci_budget_budget.budget1.time_spend_computed}"
+  }
 }
 
 data "oci_budget_budgets" "test_budgets" {
@@ -95,22 +94,21 @@ data "oci_budget_alert_rules" "test_alert_rules" {
 }
 
 output "alert_rule" {
-  value = <<EOF
-
-  budget_id = "${data.oci_budget_alert_rule.test_alert_rule.budget_id}"
-  recipients = "${data.oci_budget_alert_rule.test_alert_rule.recipients}"
-  description = "${data.oci_budget_alert_rule.test_alert_rule.description}"
-  display_name = "${data.oci_budget_alert_rule.test_alert_rule.display_name}"
-  message = "${data.oci_budget_alert_rule.test_alert_rule.message}"
-  recipients = "${data.oci_budget_alert_rule.test_alert_rule.recipients}"
-  state = "${data.oci_budget_alert_rule.test_alert_rule.state}"
-  threshold = "${data.oci_budget_alert_rule.test_alert_rule.threshold}"
-  threshold_type = "${data.oci_budget_alert_rule.test_alert_rule.threshold_type}"
-  time_created = "${data.oci_budget_alert_rule.test_alert_rule.time_created}"
-  time_updated = "${data.oci_budget_alert_rule.test_alert_rule.time_updated}"
-  type = "${data.oci_budget_alert_rule.test_alert_rule.type}"
-  version = "${data.oci_budget_alert_rule.test_alert_rule.version}"
-EOF
+  value = {
+    budget_id      = "${data.oci_budget_alert_rule.test_alert_rule.budget_id}"
+    recipients     = "${data.oci_budget_alert_rule.test_alert_rule.recipients}"
+    description    = "${data.oci_budget_alert_rule.test_alert_rule.description}"
+    display_name   = "${data.oci_budget_alert_rule.test_alert_rule.display_name}"
+    message        = "${data.oci_budget_alert_rule.test_alert_rule.message}"
+    recipients     = "${data.oci_budget_alert_rule.test_alert_rule.recipients}"
+    state          = "${data.oci_budget_alert_rule.test_alert_rule.state}"
+    threshold      = "${data.oci_budget_alert_rule.test_alert_rule.threshold}"
+    threshold_type = "${data.oci_budget_alert_rule.test_alert_rule.threshold_type}"
+    time_created   = "${data.oci_budget_alert_rule.test_alert_rule.time_created}"
+    time_updated   = "${data.oci_budget_alert_rule.test_alert_rule.time_updated}"
+    type           = "${data.oci_budget_alert_rule.test_alert_rule.type}"
+    version        = "${data.oci_budget_alert_rule.test_alert_rule.version}"
+  }
 }
 
 data "oci_budget_alert_rule" "test_alert_rule" {