+++ examples/compute/instance/variables.tf
@@ -57,7 +57,7 @@ variable "tag_namespace_description" {
 }
 
 variable "tag_namespace_name" {
-  default = "exampletagns"
+  default = "testexamples-tag-namespace"
 }
 
 variable "volume_attachment_device" {