+++ oci/core_virtual_circuit_bandwidth_shape_test.go
@@ -15,8 +15,7 @@ var (
 		"provider_service_id": Representation{repType: Required, create: `${data.oci_core_fast_connect_provider_services.test_fast_connect_provider_services.fast_connect_provider_services.0.id}`},
 	}
 
-	// @CODEGEN 07/2018: ProviderService is actually FastConnectProviderService
-	VirtualCircuitBandwidthShapeResourceDependencies = FastConnectProviderServiceResourceConfig
+	VirtualCircuitBandwidthShapeResourceConfig = generateDataSourceFromRepresentationMap("oci_core_fast_connect_provider_services", "test_fast_connect_provider_services", Required, Create, fastConnectProviderServiceDataSourceRepresentation)
 )
 
 func TestCoreVirtualCircuitBandwidthShapeResource_basic(t *testing.T) {
@@ -36,15 +35,9 @@ func TestCoreVirtualCircuitBandwidthShapeResource_basic(t *testing.T) {
 		Steps: []resource.TestStep{
 			// verify datasource
 			{
-				Config: config + `
-
-data "oci_core_fast_connect_provider_services" "test_fast_connect_provider_services" {
-	#Required
-	compartment_id = "${var.compartment_id}"
-
-}` +
+				Config: config +
 					generateDataSourceFromRepresentationMap("oci_core_virtual_circuit_bandwidth_shapes", "test_virtual_circuit_bandwidth_shapes", Required, Create, virtualCircuitBandwidthShapeDataSourceRepresentation) +
-					compartmentIdVariableStr + VirtualCircuitBandwidthShapeResourceDependencies,
+					compartmentIdVariableStr + VirtualCircuitBandwidthShapeResourceConfig,
 				Check: resource.ComposeAggregateTestCheckFunc(
 
 					resource.TestCheckResourceAttrSet(datasourceName, "virtual_circuit_bandwidth_shapes.#"),